Move types to subdirectory
