+++ src/app/configurator/configurator.component.ts
@@ -5,7 +5,7 @@ import { Observable } from 'rxjs';
 import { CardComponent } from '../shared/components/card/card.component';
 import { PageComponent } from '../shared/components/page/page.component';
 import { ConfiguratorService } from './configurator.service';
-import { DisabledSteps } from './shared/disabled-steps.type';
+import { DisabledSteps } from './shared/types/disabled-steps.type';
 
 @Component({
   selector: 'app-configurator',
