--- src/app/configurator/shared/options.type.ts
@@ -1,3 +1,0 @@
-import { ProcessedOptionsApiData } from '../../shared/services/api/types/processed-options-api-data.type';
-
-export type Options = ProcessedOptionsApiData;
