+++ src/app/configurator/shared/types/configurator-form.type.ts
