+++ src/app/configurator/shared/types/disabled-steps.type.ts
