+++ src/app/configurator/shared/types/image.type.ts
