+++ src/app/configurator/shared/types/options.type.ts
@@ -1,0 +1,3 @@
+import { ProcessedOptionsApiData } from '../../../shared/services/api/types/processed-options-api-data.type';
+
+export type Options = ProcessedOptionsApiData;
